requirements.txt
Flask==2.1.2
flask-restx==0.5.1
cassandra-driver==3.25.0
pydrive==1.3.1